fix(sendaudio): add_sound returns queue_empty threading.Event
In order to make the API more ergonomic, sendaudio.add_sound() now
returns a threading Event that activates when the send audio queue is
empty. This makes code waiting on the audio being sent cleaner, although
if multiple sounds are queued to be sent before the queue is empty, the
Event will not activate until all sounds are sent.