fix(callbacks): simplify callbacks interface
As recommended in the documentation, avoid subclass `dict` by defining
a Callbacks dataclass containing a Callback object for each callback
type.

Adds documentation for when callbacks are called, how to register
callbacks, and what parameters are passed to each callback.

Fixes examples to use the new interface. Fixes the audio-only client
example to actually receive and send audio, and adds argument parsing.